MentorMateStage1/requirements.txt
chromadb==0.5.3
langchain==0.2.6
langchain_ai21==0.1.6
langchain_community==0.2.6
langchain_core==0.2.10
langchain_groq==0.1.6
python-dotenv==1.0.1
pypdf == 4.2.0
